requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
langchain==0.3.25
pymongo==4.5.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.78.1
pydantic==2.11.4
pydantic-settings==2.9.1
pytest==7.4.2
sqlalchemy==2.0.20
sqlparse==0.4.4
python-jose[cryptography]==3.3.0
httpx==0.24.1
langchain-openai==0.3.16
python-multipart==0.0.2